# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: jupiq Control Plane API
  version: 1.1.0
  description: >-
    오프라인망 JupyterHub 통합 운영, 사용량, 정책, 승인, AI 및 MCP API.
    성공 응답은 기본적으로 data 필드를 사용하며 목록은 meta 페이지 정보를 포함한다.
    객체 범위 역할은 Hub 목록·상세·변경·삭제·테스트·동기화, 통합 사용자 목록,
    서버 목록·제어에만 적용된다. 그 밖의 관리·통계·검색·MCP API는 전역 권한이 필요하다.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      security: []
      tags: [Auth]
      operationId: localLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Error' }
...
  /auth/password:
    post:
      tags: [Auth]
      operationId: changeLocalPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password }
                new_password: { type: string, format: password, minLength: 12 }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/Error' }
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      responses:
        '200': { $ref: '#/components/responses/Success' }
    patch:
      tags: [Auth]
      operationId: updateCurrentUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name: { type: string }
                email: { type: string, format: email }
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
  /servers/{id}/{action}:
    post:
      tags: [Operations]
      operationId: controlServer
      x-access-scope: global-or-target-hub-and-department
      parameters:
        - { $ref: '#/components/parameters/ID' }
        - { name: action, in: path, required: true, schema: { type: string, enum: [start, stop, restart] } }
      responses:
        '202': { $ref: '#/components/responses/Success' }
...
  /ai/chat:
    post:
      tags: [AI]
      operationId: streamAIChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              properties:
                model: { type: string }
                messages: { type: array, items: { type: object, additionalProperties: true } }
                max_tokens: { type: integer, minimum: 1, maximum: 262144 }
                temperature: { type: number }
      responses:
        '200':
          description: OpenAI 호환 SSE 스트림
          content: { text/event-stream: { schema: { type: string } } }
...
  /profiles/{id}:
    get: { tags: [Administration], operationId: getProfile, parameters: [{ $ref: '#/components/parameters/ID' }], responses: { '200': { $ref: '#/components/responses/Success' } } }
    put: { tags: [Administration], operationId: updateProfile, parameters: [{ $ref: '#/components/parameters/ID' }], requestBody: { $ref: '#/components/requestBodies/ResourceWrite' }, responses: { '200': { $ref: '#/components/responses/Success' } } }
    delete: { tags: [Administration], operationId: deleteProfile, parameters: [{ $ref: '#/components/parameters/ID' }], responses: { '204': { description: 삭제 완료 } } }
...
  /mcp:
    post:
      tags: [AI]
      operationId: mcpStreamableHTTP
      description: /mcp의 /api/v1 별칭. initialize, ping, tools/list, tools/call 지원.
      requestBody:
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200':
          description: JSON-RPC JSON 또는 SSE 응답
          content:
            application/json: { schema: { type: object, additionalProperties: true } }
            text/event-stream: { schema: { type: string } }
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: jupiq_session }
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT-or-jqk }
  parameters:
    ID: { name: id, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
  requestBodies:
...
    ResourceWrite:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [name]
            properties:
              name: { type: string }
              status: { type: string }
              owner_user_id: { type: integer, format: int64 }
              data: { type: object, additionalProperties: true }
            additionalProperties: true
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
...
  responses:
    Success:
      description: 성공
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties: { data: {} }
...
    Error:
      description: 오류
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }